+++ extensions/robospice-okhttp-parent/robospice-okhttp-test/src/main/java/com/octo/android/robospice/okhttp/test/OkHttpSpiceTestService.java
@@ -1,0 +1,16 @@
+package com.octo.android.robospice.okhttp.test;
+
+import android.app.Application;
+
+import com.octo.android.robospice.okhttp.OkHttpSpiceService;
+import com.octo.android.robospice.persistence.CacheManager;
+import com.octo.android.robospice.persistence.exception.CacheCreationException;
+
+public class OkHttpSpiceTestService extends OkHttpSpiceService {
+
+    @Override
+    public CacheManager createCacheManager(Application application) throws CacheCreationException {
+        return new CacheManager();
+    }
+
+}